Finished CRUD User + Task Admin Manage Account Information

+++ NewsManagementSystem/Repositories/DTOs/SystemAccountDTOs/GetSystemAccountDTO.cs
@@ -2,7 +2,7 @@
 {
     public class GetSystemAccountDTO : BaseSystemAccountDTO
     {
-        public int AccountId { get; set; }
+        public short AccountId { get; set; }
         public string RoleName { get; set; } = string.Empty;
     }
 }
